+++ CHANGES.rst
@@ -5,7 +5,15 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add missing translation domain
+  [laulaz]
+
+- Add basic style for sortable hover
+  This style has disappeared in Plone 6 (>< Plone 5)
+  [laulaz]
+
+- Fix add/edit forms no-tabbing feature for Plone 6
+  [laulaz]
 
 
 1.0a1 (2021-04-19)